move spring-boot-app module content to newly created module "spring-server"+ run tests
