--- application/spring-boot-app/build.gradle
@@ -1,82 +1,0 @@
-/*
- * Copyright (c) 2020. MobilityData IO.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'org.springframework.boot' version '2.3.2.RELEASE'
-    id 'io.spring.dependency-management' version '1.0.9.RELEASE'
-    id 'java'
-    id 'war'
-}
-
-group 'org.mobilitydata'
-version '1.3.0-SNAPSHOT_web'
-
-repositories {
-    mavenCentral()
-}
-
-
-dependencies {
-    implementation(
-            project(':config'),
-            project(':react-client'),
-            project(':domain'),
-            project(':usecase'),
-    )
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    testImplementation('org.springframework.boot:spring-boot-starter-test') {
-        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
-    }
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.5.2'
-    testImplementation 'org.hamcrest:hamcrest:2.2'
-    testImplementation 'org.mockito:mockito-core:3.3.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.5.2'
-
-    testImplementation('org.springframework.boot:spring-boot-starter-test') {
-        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
-    }
-    providedRuntime 'org.springframework.boot:spring-boot-starter-tomcat'
-}
-
-bootWar {
-    print(System.getProperty("user.dir"))
-    task copyDocs(type: War) {
-        from ("/home/runner/work/gtfs-validator/gtfs-validator/build/libs/gtfs-validator.jar") { // "/home/runner/work/gtfs-validator/gtfs-validator/react-client/build"
-            into "./META-INF/resources"
-        }
-    }
-    // Change the JAR name from spring-boot-app to gtfs-validator
-    archiveBaseName = rootProject.name
-    // Remove "-all" from the end of the JAR file name
-    archiveClassifier = null
-    // Add "v" to version to match previous releases
-    archiveVersion = "v" + project.version
-}
-
-test {
-    // Enable JUnit 5 (Gradle 4.6+).
-    useJUnitPlatform()
-
-    // Always run tests, even when nothing changed.
-    dependsOn 'cleanTest'
-
-    // Show test results.
-    testLogging {
-        events "passed", "skipped", "failed"
-    }
-}
