+++ application/web-app/spring-server/src/test/java/org/mobilitydata/gtfsvalidator/web/ServiceManagerTest.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator;
+package org.mobilitydata.gtfsvalidator.web;
 
 import org.junit.jupiter.api.Test;
