+++ settings.gradle
@@ -19,8 +19,6 @@ include 'adapter:tree'
 findProject(':adapter:tree')?.name = 'tree'
 include 'adapter:utils'
 findProject(':adapter:utils')?.name = 'utils'
-include 'application:spring-boot-app'
-findProject(':application:spring-boot-app')?.name = 'spring-boot-app'
 include 'application:web-app'
 findProject(':application:web-app')?.name = 'web-app'
 include 'application:web-app:spring-server'
